SGCZ-ChillZone/requirements.txt
bcrypt==4.2.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.0.3
Flask-WTF==1.2.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
packaging==25.0
pluggy==1.6.0
pytest==8.3.2
pytest-mock==3.14.0
python-dotenv==1.0.1
Werkzeug==3.1.3
WTForms==3.1.2
